remove init from simulation results; start of adding ensemble() functionality; changed default to only_stocks = TRUE; created Julia package; added macros to model_setup; remove the option to pass a string as a unit definition to convert_u so that there is no need for a global variable unit_context (which was necessary as Unitful.uparse() does not recognize custom defined units)
