+++ NAMESPACE
@@ -2,6 +2,7 @@
 
 S3method(as.data.frame,sdbuildR_sim)
 S3method(as.data.frame,sdbuildR_xmile)
+S3method(plot,sdbuildR_ensemble)
 S3method(plot,sdbuildR_sim)
 S3method(plot,sdbuildR_xmile)
 S3method(summary,sdbuildR_xmile)
@@ -19,6 +20,7 @@ export(debugger)
 export(delay)
 export(delayN)
 export(drop_u)
+export(ensemble)
 export(expit)
 export(get_build_code)
 export(get_regex_time_units)
