+++ inst/sdbuildRUtils/src/custom_func.jl
@@ -1,0 +1,313 @@
+# custom_func
+module custom_func
+using Unitful
+using DataInterpolations
+using Distributions
+using ..unit_func: convert_u
+
+# Extrapolation function
+function itp(x, y; method = "linear", extrapolation = "nearest")
+
+  # Ensure y is sorted along x
+  idx = sortperm(x)
+  x = x[idx]
+  y = y[idx]
+
+  # Extrapolation rule: What happens outside of defined values?
+  # Rule "NA": return NaN; Rule "nearest": return closest value
+  rule_method = ifelse(extrapolation == "NA", DataInterpolations.ExtrapolationType.None, ifelse(extrapolation == "nearest", DataInterpolations.ExtrapolationType.Constant, extrapolation))
+
+  if method == "constant"
+      func = DataInterpolations.ConstantInterpolation(y, x; extrapolation = rule_method) # notice order of x and y
+  elseif method == "linear"
+      func = DataInterpolations.LinearInterpolation(y, x; extrapolation = rule_method)
+  end
+
+  return(func)
+end
+
+function ramp(times, time_units, start, finish, height = 1.0)
+
+    @assert start < finish "The finish time of the ramp cannot be before the start time. To specify a decreasing ramp, set the height to a negative value."
+
+    # If times has units, but the ramp times don't, convert them to the same units
+    if eltype(times) <: Unitful.Quantity
+        if !(eltype(start) <: Unitful.Quantity)
+            start = convert_u(start, time_units)
+        end
+        if !(eltype(finish) <: Unitful.Quantity)
+            finish = convert_u(finish, time_units)
+        end
+    else
+        # If times does not have units, but start does, convert the ramp times to the same units as time_units
+        if eltype(start) <: Unitful.Quantity
+            start = Unitful.ustrip(convert_u(start, time_units))
+        end
+        if eltype(finish) <: Unitful.Quantity
+            finish = Unitful.ustrip(convert_u(finish, time_units))
+        end
+    end
+
+
+    # Ensure start_h_ramp and height are both of the same type
+    start_h_ramp = 0.0
+    if !(eltype(height) <: Unitful.Quantity)
+        height = convert_u(height, Unitful.unit(start_h_ramp))
+        add_y = convert_u(0.0, Unitful.unit(start_h_ramp))
+    elseif eltype(height) <: Unitful.Quantity
+        start_h_ramp = convert_u(start_h_ramp, Unitful.unit(height))
+        add_y = convert_u(0.0, Unitful.unit(height))
+    else
+        add_y = 0.0
+    end
+
+    x = [start, finish]
+    y = [start_h_ramp, height]
+
+    # If the ramp is after the start time, add a zero at the start
+    if start > first(times)
+        x = [first(times); x]
+        y = [add_y; y]
+    end
+
+    func = itp(x, y, method = "linear", extrapolation = "nearest")
+
+    return(func)
+end 
+
+# Make step signal
+function make_step(times, time_units, start, height = 1.0)
+
+    # If times has units, but the ramp times don't, convert them to the same units
+    if eltype(times) <: Unitful.Quantity
+        if !(eltype(start) <: Unitful.Quantity)
+            start = convert_u(start, time_units)
+        end
+    else
+        # If times does not have units, but start does, convert the ramp times to the same units as time_units
+        if eltype(start) <: Unitful.Quantity
+            start = Unitful.ustrip(convert_u(start, time_units))
+        end
+    end
+
+    if eltype(height) <: Unitful.Quantity
+      add_y = convert_u(0.0, Unitful.unit(height))
+    else
+      add_y = 0.0
+    end
+
+    x = [start, times[2]]
+    y = [height, height]
+
+    # If the step is after the start time, add a zero at the start
+    if start > first(times)
+        x = [first(times); x]
+        y = [add_y; y]
+    end
+
+    func = itp(x, y, method = "constant", extrapolation = "nearest")
+
+    return(func)
+end
+
+    
+
+# Make pulse signal
+function pulse(times, time_units, start, height = 1.0, width = 1.0 * time_units, repeat_interval = nothing)
+    # If times has units, but the pulse times don't, convert them to the same units
+    if eltype(times) <: Unitful.Quantity
+        if !(eltype(start) <: Unitful.Quantity)
+            start = convert_u(start, time_units)
+        end
+        if !(eltype(width) <: Unitful.Quantity)
+            width = convert_u(width, time_units)
+        end
+        if (!isnothing(repeat_interval) && !(eltype(repeat_interval) <: Unitful.Quantity))
+            repeat_interval = convert_u(repeat_interval, time_units)
+        end
+    else
+        # If times does not have units, but start does, convert the pulse times to the same units as time_units
+        if eltype(start) <: Unitful.Quantity
+            start = Unitful.ustrip(convert_u(start, time_units))
+        end
+        if eltype(width) <: Unitful.Quantity
+            width = Unitful.ustrip(convert_u(width, time_units))
+        end
+        if (!isnothing(repeat_interval) && eltype(repeat_interval) <: Unitful.Quantity)
+            repeat_interval = Unitful.ustrip(convert_u(repeat_interval, time_units))
+        end
+    end
+
+    # Define start and end times of pulses
+    last_time = last(times)
+    # If no repeats, set end of pulse to after end time
+    step_size = isnothing(repeat_interval) ? last_time * 2 : repeat_interval
+    start_ts = collect(start:step_size:last_time)
+    end_ts = start_ts .+ width
+
+    # Build signal as vectors of times and y-values
+    signal_times = [start_ts; end_ts]
+    signal_y = [fill(height, length(start_ts)); fill(0, length(end_ts))]
+
+    if eltype(height) <: Unitful.Quantity
+      add_y = convert_u(0.0, Unitful.unit(height))
+    else
+      add_y = 0.0
+    end
+
+    # If the first pulse is after the start time, add a zero at the start
+    if minimum(start_ts) > first(times)
+        signal_times = [first(times); signal_times]
+        signal_y = [add_y; signal_y]
+    end
+
+    # If the last pulse doesn't cover the end, add a zero at the end
+    # (I don't fully understand why this is necessary, but otherwise it gives incorrect results with repeat_interval <= 0)
+    if maximum(end_ts) < last_time
+        signal_times = [signal_times; last_time]
+        signal_y = [signal_y; add_y]
+    end
+
+    # Sort by time
+    perm = sortperm(signal_times)
+    x = signal_times[perm]
+    y = signal_y[perm]
+    func = itp(x, y, method = "constant", extrapolation = "nearest")
+
+    return(func)
+end
+
+# Create seasonal wave 
+function seasonal(times, dt, period = u"1yr", shift = u"0yr")
+
+    @assert period > 0 "The period of the seasonal wave must be greater than 0."
+
+    time_vec = times[1]:dt:times[2]
+    phase = 2 * pi .* (time_vec .- shift) ./ period  # π radians
+    y = cos.(phase)
+    func = itp(time_vec, y, method = "linear", extrapolation = "nearest")
+
+    return(func)
+end
+
+# Convert Insight Maker's Round() function to R
+# Difference: in Insight Maker, Round(.5) = 1; in R, round(.5) = 0; in julia, round(.5) = 0.0
+function IM_round(x::Real, digits::Int=0)
+    # Compute the fractional part after scaling by 10^digits
+    scaled_x = x * 10.0^digits
+    frac = scaled_x % 1
+    # Check if fractional part is exactly 0.5 or -0.5
+    if abs(frac) == 0.5
+        return ceil(scaled_x) / 10.0^digits
+    else
+        return round_(scaled_x, digits=0) / 10.0^digits
+    end
+end
+
+# Logit function
+function logit(p)
+    return log(p / (1 - p))
+end
+
+# Expit function
+function expit(x)
+    return 1 / (1+exp(-x))
+end
+
+function logistic(x, slope=1.0, midpoint=0.0, upper = 1.0)
+    @assert isfinite(slope) && isfinite(midpoint) && isfinite(upper) "slope, midpoint, and upper must be numeric"
+    upper / (1 + exp(-slope * (x - midpoint)))
+end
+
+
+# Prevent non-negativity (below zero)
+# Scalar case: non-unitful types
+nonnegative(x::Real) = max(0.0, x)
+
+# Scalar case: Unitful.Quantity
+nonnegative(x::Unitful.Quantity) = max(0.0, Unitful.ustrip(x)) * Unitful.unit(x)
+
+# Array case: non-unitful elements
+nonnegative(x::AbstractArray{<:Real}) = max.(0.0, x)
+
+# Array case: Unitful.Quantity elements
+nonnegative(x::AbstractArray{<:Unitful.Quantity}) = max.(0.0, Unitful.ustrip.(x)) .* Unitful.unit.(x)
+
+# Generate random boolean value, equivalent of RandBoolean() in Insight Maker
+function rbool(p)
+    return rand() < p
+end
+
+function rdist(a::Vector{T}, b::Vector{<:Real}) where T
+    # Check lengths match
+    if length(a) != length(b)
+        throw(ArgumentError("Length of a and b must match"))
+    end
+    # Normalize probabilities
+    b_sum = sum(b)
+    if b_sum <= 0
+        throw(ArgumentError("Sum of probabilities must be positive"))
+    end
+    b_normalized = b / b_sum
+    # Sample using Categorical
+    return a[rand(Distributions.Categorical(b_normalized))]
+end
+
+function indexof(haystack, needle)
+    if isa(haystack, AbstractString) && isa(needle, AbstractString)
+        pos = findfirst(needle, haystack)
+        return isnothing(pos) ? 0 : first(pos)
+    else
+        pos = findfirst(==(needle), haystack)
+        return isnothing(pos) ? 0 : pos
+    end
+end
+
+function IM_contains(haystack, needle)
+    if isa(haystack, AbstractString) && isa(needle, AbstractString)
+        return occursin(needle, haystack)
+    else
+        return needle in haystack
+    end
+end
+
+function substr_i(string::AbstractString, idxs::Union{Int, Vector{Int}})
+    chars = collect(string)
+    return join(chars[idxs])
+end
+
+function IM_filter(y::Vector{T}, condition_func::Function) where T
+    names_y = string.(1:length(y))
+    result = Dict{String,T}()
+    for (key, val) in zip(names_y, y)
+        if condition_func(val, key)
+            result[key] = val
+        end
+    end
+    return collect(values(result)), collect(keys(result))
+end
+
+
+#round_(x::Unitful.Quantity) = round(Unitful.ustrip.(x)) * Unitful.unit(x)
+#round_(x::Unitful.Quantity, digits::Int) = round(Unitful.ustrip.(x), digits=digits) * Unitful.unit(x)
+#round_(x::Unitful.Quantity; digits::Int) = round(Unitful.ustrip.(x), digits=digits) * Unitful.unit(x)
+#round_(x::Unitful.Quantity, digits::Float64) = round(Unitful.ustrip.(x), digits=round(digits)) * Unitful.unit(x)
+#round_(x::Unitful.Quantity; digits::Float64) = round(Unitful.ustrip.(x), digits=round(digits)) * Unitful.unit(x)
+
+#round_(x) = round(x)
+
+round_(x, digits::Real) = round(x, digits=round(Int, digits))
+
+round_(x; digits::Real=0) = round(x, digits=round(Int, digits))
+
+#round_(x::Unitful.Quantity) = round(Unitful.ustrip(x)) * Unitful.unit(x)
+
+round_(x::Unitful.Quantity, digits::Real) = round(Unitful.ustrip(x), digits=round(Int, digits)) * Unitful.unit(x)
+
+round_(x::Unitful.Quantity; digits::Real=0) = round(Unitful.ustrip(x), digits=round(Int, digits)) * Unitful.unit(x)
+
+# Define the operator ⊕ for the modulus
+function ⊕(x, y)
+    return mod(x, y)
+end
+end
