+++ man/ensemble.Rd
@@ -1,0 +1,59 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/compile.R
+\name{ensemble}
+\alias{ensemble}
+\title{Create ensemble of simulations}
+\usage{
+ensemble(
+  sfm,
+  n = 10,
+  summary = TRUE,
+  qs = c(0.025, 0.975),
+  return_sims = FALSE,
+  range = NULL,
+  keep_nonnegative_flow = TRUE,
+  keep_nonnegative_stock = FALSE,
+  keep_unit = TRUE,
+  only_stocks = FALSE,
+  verbose = FALSE,
+  debug = FALSE
+)
+}
+\arguments{
+\item{sfm}{Stock-and-flow model, object of class sdbuildR_xmile.}
+
+\item{n}{Number of simulations to run in the ensemble. Defaults to 10.}
+
+\item{summary}{If TRUE, summarize the ensemble simulation results using their mean, variance, median, and quantiles. Defaults to TRUE.}
+
+\item{qs}{Quantiles to calculate in the summary, e.g. c(0.025, 0.975)}
+
+\item{return_sims}{If TRUE, return the individual simulations in the ensemble. Defaults to FALSE.}
+
+\item{range}{List of ranges to vary parameters in the ensemble. If specified, \code{n} will be the length of that range. All ranges have to be of the same length. Defaults to NULL.}
+
+\item{keep_nonnegative_flow}{If TRUE, keeps original non-negativity setting of flows. Defaults to TRUE.}
+
+\item{keep_nonnegative_stock}{If TRUE, keeps original non-negativity setting of stocks Defaults to TRUE.}
+
+\item{keep_unit}{If TRUE, keeps units of variables. Defaults to TRUE.}
+
+\item{only_stocks}{If TRUE, only save stocks. If FALSE, auxiliaries and flows are saved using a callback function. Only applies if language is set to "Julia" in sim_specs() and no delay functions are used. Defaults to FALSE.}
+
+\item{verbose}{If TRUE, update on progress. Defaults to FALSE.}
+
+\item{debug}{If TRUE, print output for debugging. Defaults to FALSE.}
+}
+\value{
+Object of class sdbuildR_ensemble, which is a list containing:
+\describe{
+\item{success}{If TRUE, simulation was successful. If FALSE, simulation failed.}
+\item{filepath_sim}{Filepath to the simulation results in .ser format.}
+\item{script}{Julia script used for the ensemble simulation.}
+\item{duration}{Duration of the simulation in seconds.}
+\item{...}{Other parameters passed to ensemble}
+}
+}
+\description{
+Create ensemble of simulations
+}
