+++ man/get_func_julia.Rd
@@ -1,0 +1,14 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/get_func_julia.R
+\name{get_func_julia}
+\alias{get_func_julia}
+\title{Customary functions written in Julia}
+\usage{
+get_func_julia()
+}
+\value{
+List with Julia code
+}
+\description{
+Customary functions written in Julia
+}
